backend/pyproject.toml
[project]
name = "django-nextjs"
version = "0.1.0"
description = "Full-stack practice project using Django Ninja REST API with Next.js frontend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.4.3",
    "django-unfold>=0.67.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
]
